moved weights by base and cell to separate folder
