+++ include/deal.II-qc/core/qc.h
@@ -28,7 +28,6 @@ namespace LA
 }
 
 #include <deal.II-qc/atom/atom_handler.h>
-#include <deal.II-qc/atom/cluster_weights.h>
 #include <deal.II-qc/configure/configure_qc.h>
 #include <deal.II-qc/potentials/pair_lj_cut.h>
 #include <deal.II-qc/potentials/pair_coul_wolf.h>
