+++ source/atom/sampling/cluster_weights_by_cell.cc
@@ -1,5 +1,5 @@
 
-#include <deal.II-qc/atom/cluster_weights.h>
+#include <deal.II-qc/atom/sampling/cluster_weights_by_cell.h>
 
 namespace dealiiqc
 {
@@ -8,35 +8,6 @@ namespace dealiiqc
   {
 
 
-
-    //------------------------------------------------------------------------//
-    // WeightsByBase
-
-    template <int dim>
-    WeightsByBase<dim>::WeightsByBase (const double &cluster_radius,
-                                       const double &maximum_cutoff_radius)
-      :
-      cluster_radius(cluster_radius),
-      maximum_cutoff_radius(maximum_cutoff_radius)
-    {}
-
-
-
-    template <int dim>
-    WeightsByBase<dim>::~WeightsByBase()
-    {}
-
-
-
-    // Instantiations
-    template class WeightsByBase<1>;
-    template class WeightsByBase<2>;
-    template class WeightsByBase<3>;
-
-
-
-    //------------------------------------------------------------------------//
-    // WeightsByCell
 
     template <int dim>
     WeightsByCell<dim>::WeightsByCell (const double &cluster_radius,
